pyproject.toml
[project]
name = "ia-vscode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
